requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=8.3.0
matplotlib>=3.5.0
wandb>=0.15.0
jupyter>=1.0.0
opencv-python>=4.5.0
albumentations>=1.3.0
tqdm>=4.64.0
scikit-image>=0.19.0
PyYAML>=6.0 